chore: refactor the models.py file and break it down into individual files
