+++ cashu/core/models/__init__.py
@@ -1,0 +1,105 @@
+from .blind_auth import PostAuthBlindMintRequest, PostAuthBlindMintResponse
+from .check import (
+    CheckFeesRequest_deprecated,
+    CheckFeesResponse_deprecated,
+    CheckSpendableRequest_deprecated,
+    CheckSpendableResponse_deprecated,
+    PostCheckStateRequest,
+    PostCheckStateResponse,
+)
+from .info import (
+    GetInfoResponse,
+    GetInfoResponse_deprecated,
+    MeltMethodSetting,
+    MintInfoContact,
+    MintInfoProtectedEndpoint,
+    MintMethodBolt11OptionSetting,
+    MintMethodSetting,
+    Nut15MppSupport,
+)
+from .keys import (
+    KeysetsResponse,
+    KeysetsResponse_deprecated,
+    KeysetsResponseKeyset,
+    KeysResponse,
+    KeysResponse_deprecated,
+    KeysResponseKeyset,
+)
+from .melt import (
+    PostMeltRequest,
+    PostMeltRequest_deprecated,
+    PostMeltResponse_deprecated,
+)
+from .melt_quote import (
+    PostMeltQuoteRequest,
+    PostMeltQuoteResponse,
+    PostMeltRequestOptionMpp,
+    PostMeltRequestOptions,
+)
+from .mint import (
+    GetMintResponse_deprecated,
+    PostMintRequest,
+    PostMintRequest_deprecated,
+    PostMintResponse,
+    PostMintResponse_deprecated,
+)
+from .mint_quote import PostMintQuoteRequest, PostMintQuoteResponse
+from .restore import (
+    PostRestoreRequest,
+    PostRestoreRequest_Deprecated,
+    PostRestoreResponse,
+)
+from .swap import (
+    PostSwapRequest,
+    PostSwapRequest_Deprecated,
+    PostSwapResponse,
+    PostSwapResponse_Deprecated,
+    PostSwapResponse_Very_Deprecated,
+)
+
+__all__ = [
+    "PostAuthBlindMintRequest",
+    "PostAuthBlindMintResponse",
+    "CheckFeesRequest_deprecated",
+    "CheckFeesResponse_deprecated",
+    "CheckSpendableRequest_deprecated",
+    "CheckSpendableResponse_deprecated",
+    "PostCheckStateRequest",
+    "PostCheckStateResponse",
+    "GetInfoResponse",
+    "GetInfoResponse_deprecated",
+    "MeltMethodSetting",
+    "MintInfoContact",
+    "MintInfoProtectedEndpoint",
+    "MintMethodBolt11OptionSetting",
+    "MintMethodSetting",
+    "Nut15MppSupport",
+    "KeysResponse",
+    "KeysResponse_deprecated",
+    "KeysResponseKeyset",
+    "KeysetsResponse",
+    "KeysetsResponse_deprecated",
+    "KeysetsResponseKeyset",
+    "PostMeltRequest",
+    "PostMeltRequest_deprecated",
+    "PostMeltResponse_deprecated",
+    "PostMeltQuoteRequest",
+    "PostMeltQuoteResponse",
+    "PostMeltRequestOptionMpp",
+    "PostMeltRequestOptions",
+    "GetMintResponse_deprecated",
+    "PostMintRequest",
+    "PostMintRequest_deprecated",
+    "PostMintResponse",
+    "PostMintResponse_deprecated",
+    "PostMintQuoteRequest",
+    "PostMintQuoteResponse",
+    "PostRestoreRequest",
+    "PostRestoreRequest_Deprecated",
+    "PostRestoreResponse",
+    "PostSwapRequest",
+    "PostSwapRequest_Deprecated",
+    "PostSwapResponse",
+    "PostSwapResponse_Deprecated",
+    "PostSwapResponse_Very_Deprecated",
+]
